pyproject.toml
[tool.poetry]
name = "orca_ui"
version = "0.1.0"
description = "Orca Frontend"
authors = ["kamal bhatt <kamal.bhatt@stordis.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.2.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
